Extra flagging for difficult days
Looked like mfcal went a bit crazy when there was strong RFI in the secondary spectrum and when MFCAL was used with a defined model and a time dependent bandpass.

An initial round of flagging seems to have helped for 2018-01-08. Rerunning with flagging on the mosaic and also applying extra flagging for 2018-05-31.